fix: Add Combined End-to-End tests for ASL recognition, Quiz, and placeholder features
